backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic>=2.12.0
pydantic-settings>=2.5.0
sqlalchemy>=2.0.36
psycopg[binary]==3.2.13
alembic>=1.13.2
celery==5.4.0
redis==5.0.7
python-multipart==0.0.9
boto3==1.34.138
structlog==24.2.0
httpx==0.27.0
python-dotenv>=1.0.0
